pyproject.toml
[tool.poetry]
name = "meme lord"
version = "0.1.0"
description = ""
authors = ["Nemika-Haj <mmnemijus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
disnake = "^2.7.0"
requests = "^2.28.1"
Pillow = "9.0.0"
moviepy = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
